ziprepack improved

+++ ziprepack.sh
@@ -10,15 +10,12 @@ TEMP=/var/www/chemax/downloads/otherstemp
 COUNTER=0
 PASSWORD=qwertyqwerty
 
-/bin/echo `date +%F" "%T`" -----------------------ziprepack started-------------------------" > /var/log/zipencrypt.log
+/bin/echo `date +%F" "%T`" -----------------------ziprepack started-------------------------" >> /var/log/zipencrypt.log
 for FILE in $(ls $SOURCE); do
-        [[ "$FILE" == *.zip ]] || {
-                echo "$FILE is not a ZIP archive" >>/var/log/zipencrypt.log
-                continue
-        }
+        [[ "$FILE" == *.zip ]] || continue
         FILERAR=`echo $FILE | sed -e 's/\.zip/\.rar/g'`
         /usr/bin/nice -n 5 /usr/bin/unzip -q $SOURCE/$FILE -d $TEMP >>/var/log/zipencrypt.log 2>&1
-        /usr/bin/nice -n 5 /usr/bin/rar a -ep -inul -hp$PASSWORD $ENCRYPTED/$FILERAR $TEMP/* >>/var/log/zipencrypt.log 2>&1
+        /usr/bin/nice -n 5 /usr/bin/rar a -ep1 -inul -hp$PASSWORD $ENCRYPTED/$FILERAR $TEMP/* >>/var/log/zipencrypt.log 2>&1
         /usr/bin/nice -n 5 /usr/bin/zip -jq $TARGET/$FILE $ENCRYPTED/$FILERAR >>/var/log/zipencrypt.log 2>&1
         /bin/rm -rf $TEMP/* >/dev/null 2>&1
         /bin/rm -rf $TEMP/.* >/dev/null 2>&1
